check exists befor unlink

+++ app/Controllers/ToolsEstosController.php
@@ -177,8 +177,12 @@ class ToolsEstosController extends BaseController
         $uploadFile->move(WRITEPATH . 'uploads/temp', 'kalahari.json', true);
 
         service('settings')->set('App.lastKalahariImportReceiver', auth()->user()->email);
+        service('settings')->set('App.lastKalahariImport', date('d.m.Y - H:i:s'));
         
-        unlink(WRITEPATH . 'export/telefonliste.csv');
+        if(file_exists(WRITEPATH . 'export/telefonliste.csv')){
+            unlink(WRITEPATH . 'export/telefonliste.csv');
+        }
+        
 
         return redirect()->route('estos.index')->with('message', 'Die Datei wurde hochgeladen und wird in den nächsten Minuten im Hintergrund verarbeitet. Du erhälst eine E-Mail sobald der Import abgeschlossen ist.');
     }
